dataloader: correct return of __iter__ type to FloatTensor

+++ data/dataloader.py
@@ -49,7 +49,7 @@ class DataLoader:
             next_datapoint = _datapoint_to_tensor_including_channel(self.dataset[index])
             _append_batch_by_datapoint(next_datapoint, batch)
             if len(batch) == self.batch_size:
-                yield batch.inputs, batch.labels
+                yield batch.inputs.float(), batch.labels.float()
                 batch_id += 1
                 batch = Batch(batch_id=batch_id)
